Add debug make target, add forward LABELJMP, more
Labels in elf64.h are now handled with the struct
ba_IMLabel, which has an array of code addresses
where a jump to the label is supposed to occur,
which works similarly how DATASGMT is handled. It
also has an array of sizes in code of the RIP-
relative address of the label, in the jump
instruction, so that when these addresses are
filled up, it is known whether to use a 1 byte or
4 byte offset. The other tweaks are pretty much
just related to memory management in elf64.h.