Fix bug in assigning to dereferencing lists
Prior to the fix, in an assignment to a
dereferencing list, the address of the rhs
expression would not actually be dereferenced,
meaning garbage would be assigned to the lhs
instead of the intended value.

For example, the following code would output a
garbage number, instead of always the expected 1.

u8[] buf = { 1 };
u8 c = garbage;
c = [buf,0];
write c; "\n";