Rewrite lexer, add f-strings, remove write keyword
F-strings (formatted strings) are a new method of
writing strings with variable interpolation to a
file or string. The addition of f-strings also
means the removal of the `write` keyword. A full
description of the features of formatted strings
added so far can be found in MANUAL.md.

Also:
Change the terminator of multiline comments from
"}#" to "#}".
Add 8-bit hexadecimal escape codes
Add '\a' escape code (bell character, aequivalent
to '\x07') and '\e' escape code (escape,
aequivalent to '\x1b')
Add some additional opcodes
Rewrite all tests to be functional with these new
changes